+++ src/main/java/com/awoodcock/armory/service/ar/BarrelService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +29,9 @@ public class BarrelService {
     }
 
     // READ
-    public Optional<Barrel> getBarrelById (int id) {
-        return barrelRepository.findById(id);
+    public Barrel getBarrelById (int id) {
+        Optional<Barrel> barrel = barrelRepository.findById(id);
+        return barrel.orElse(null);
     }
 
     public List<Barrel> getAllBarrels() {
@@ -38,19 +40,49 @@ public class BarrelService {
 
     // UPDATE
 
+    public Barrel updateBarrelById(int id, Barrel barrel) {
+        Optional<Barrel> barrelOptional = barrelRepository.findById(id);
+        if (barrelOptional.isPresent()) {
+            Barrel originalBarrel = barrelOptional.get();
 
-    // name
+            // name
+            if (Objects.nonNull(barrel.getName()) && !"".equalsIgnoreCase(barrel.getName())) {
+                originalBarrel.setName(barrel.getName());
+            }
 
-    // description
+            // description
+            if (Objects.nonNull(barrel.getDescription()) && !"".equalsIgnoreCase(barrel.getName())) {
+                originalBarrel.setDescription(barrel.getDescription());
+            }
 
-    // price
+            // price
+            if (Objects.nonNull(barrel.getPrice()) && barrel.getPrice() != 0) {
+                originalBarrel.setPrice(barrel.getPrice());
+            }
 
-    // weight
+            // weight
+            if (Objects.nonNull(barrel.getWeight()) && barrel.getWeight() != 0) {
+                originalBarrel.setWeight(barrel.getWeight());
+            }
 
-    //
+            // barrelLength
+            if (Objects.nonNull(barrel.getLength()) && barrel.getLength() != 0) {
+                originalBarrel.setLength(barrel.getLength());
+            }
+
+            return barrelRepository.save(originalBarrel);
+        }
+
+        return null;
+    }
+
 
     // DELETE
-    public void deleteBarrel(int id) {
-        barrelRepository.deleteById(id);
+    public String deleteBarrel(int id) {
+        if (barrelRepository.findById(id).isPresent()) {
+            barrelRepository.deleteById(id);
+            return "Barell successfully deleted.";
+        }
+        return "Barrel not found within database.";
     }
 }
