+++ src/main/java/com/awoodcock/armory/service/ar/ChargingHandleService.java
@@ -5,7 +5,9 @@ import com.awoodcock.armory.models.ar.ChargingHandle;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -28,8 +30,9 @@ public class ChargingHandleService {
     }
 
     // READ
-    public Optional<ChargingHandle> getChargingHandleById (int id) {
-        return chargingHandleRepository.findById(id);
+    public ChargingHandle getChargingHandleById (int id) {
+        Optional<ChargingHandle> chargingHandle = chargingHandleRepository.findById(id);
+        return chargingHandle.orElse(null);
     }
 
     public List<ChargingHandle> getAllChargingHandles() {
@@ -37,22 +40,53 @@ public class ChargingHandleService {
     }
 
     // UPDATE
+    public ChargingHandle updateChargingHandleById (int id, ChargingHandle chargingHandle) {
+        Optional<ChargingHandle> chargingHandleOptional = chargingHandleRepository.findById(id);
 
+        if (chargingHandleOptional.isPresent()) {
+            ChargingHandle originalChargingHandle = chargingHandleOptional.get();
 
+            // name
+            if (Objects.nonNull(chargingHandle.getName()) && !"".equalsIgnoreCase(chargingHandle.getName())) {
+                originalChargingHandle.setName(chargingHandle.getName());
+            }
 
-    // name
+            // description
+            if (Objects.nonNull(chargingHandle.getDescription()) && !"".equalsIgnoreCase(chargingHandle.getName())) {
+                originalChargingHandle.setDescription(chargingHandle.getDescription());
+            }
 
-    // description
+            // price
+            if (Objects.nonNull(chargingHandle.getPrice()) && chargingHandle.getPrice() != 0) {
+                originalChargingHandle.setPrice(chargingHandle.getPrice());
+            }
 
-    // price
+            // weight
+            if (Objects.nonNull(chargingHandle.getWeight()) && chargingHandle.getWeight() != 0) {
+                originalChargingHandle.setWeight(chargingHandle.getWeight());
+            }
 
-    // weight
+            // color
+            if (Objects.nonNull(chargingHandle.getColor()) && !"".equalsIgnoreCase(chargingHandle.getColor())) {
+                originalChargingHandle.setColor(chargingHandle.getColor());
+            }
 
-    //
+            // isAmbi
+            if (Objects.nonNull(chargingHandle.isAbmi())) {
+                originalChargingHandle.setAbmi(chargingHandle.isAbmi());
+            }
 
+            return chargingHandleRepository.save(originalChargingHandle);
+        }
+        return null;
+    }
 
     // DELETE
-    public void deleteChargingHandle(int id) {
-        chargingHandleRepository.deleteById(id);
+    public String deleteChargingHandle(int id) {
+        if (chargingHandleRepository.findById(id).isPresent()) {
+            chargingHandleRepository.deleteById(id);
+            return "Charging Handle successfully deleted.";
+        }
+        return "Charging Handle not found within database.";
     }
 }
